receiver/requirements.txt
envelope==1.5.3
python-gnupg==0.4.7